% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_hunspell.R
\name{stem_hunspell}
\alias{stem_hunspell}
\title{Stem words using Hunspell}
\usage{
stem_hunspell(words, complete = TRUE)
}
\arguments{
\item{words}{\code{character} vector of Dutch words}

\item{complete}{should complete stems be returned \code{TRUE}}
}
\description{
Stem words using Hunspell
}

