% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_dicpath_nl.R
\name{set_dicpath_nl}
\alias{set_dicpath_nl}
\title{Adds Dutch dictionary to Hunspell dicpath}
\usage{
set_dicpath_nl()
}
\description{
Sets the environment variable \code{DICPATH} (if unset) to the package
directory.
}

